Clean-up util folder: make all singular form; move lose files to it
