+++ packages/noir-ethereum-api/src/noir/oracles/accountOracles.ts
@@ -1,8 +1,8 @@
 import { type ForeignCallOutput } from '@noir-lang/noir_js';
 import { fromRlp, type GetProofReturnType, type Hex, isHex, type PublicClient } from 'viem';
-import { assert } from '../../assert.js';
+import { assert } from '../../util/assert.js';
 import { decodeField, decodeHexAddress, encodeField, encodeHex } from '../encode.js';
-import { padArray } from '../../arrays.js';
+import { padArray } from '../../util/array.js';
 import { NoirArguments } from './oracles.js';
 
 const PROOF_ONE_LEVEL_LENGTH = 532;
