+++ packages/noir-ethereum-api/src/noir/oracles/headerOracle.ts
@@ -1,9 +1,9 @@
 import { type ForeignCallOutput } from '@noir-lang/noir_js';
 import { type BlockHeader, blockToHeader, headerToRlp } from '../../ethereum/blockHeader.js';
 import { decodeField, encodeField, encodeHex } from '../encode.js';
-import { padArray } from '../../arrays.js';
+import { padArray } from '../../util/array.js';
 import { type GetBlockReturnType, hexToBytes, isHex, keccak256, type PublicClient } from 'viem';
-import { assert } from '../../assert.js';
+import { assert } from '../../util/assert.js';
 import { NoirArguments } from './oracles.js';
 
 export const MAX_HEADER_RLP_SIZE = 708;
