+++ packages/noir-ethereum-api/src/util/assert.ts
