+++ packages/noir-ethereum-api/src/util/file.ts
