+++ packages/noir-ethereum-api/src/util/function.ts
