+++ packages/noir-ethereum-api/src/util/json-bigint.ts
