+++ packages/noir-ethereum-api/src/util/mock.ts
