+++ packages/noir-ethereum-api/src/util/object.ts
