+++ packages/noir-ethereum-api/src/util/string.ts
