+++ packages/noir-ethereum-api/test/ethereum/recordingClient.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { Call, createRecordingClient } from '../../src/ethereum/recordingClient.js';
-import { withTempFile, writeObject } from '../../src/utils/file.js';
+import { withTempFile, writeObject } from '../../src/util/file.js';
 import { createMockClient } from '../../src/ethereum/mockClient.js';
 import { GetBlockParameters, GetProofParameters, Hex, PublicClient } from 'viem';
